Fix an overflow condition in the mdns refresh timer
This removes the static_cast<int> from the mDns refresh timers which can
overflow given a large-enough ttl. TimeDelta already provides clamped
type conversions (to int_64t) to prevent overflow so the casting is
unnecessary.

Bug: 464515864
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7674254
Commit-Queue: Patrick Meenan <pmeenan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1602180}

+++ net/dns/mdns_client_impl.cc
@@ -636,14 +636,12 @@ void MDnsListenerImpl::ScheduleNextRefresh() {
   // be canceled and rescheduled if the record's TTL is updated due to a
   // response being received.
   base::Time next_refresh1 =
-      last_update_ +
-      base::Milliseconds(static_cast<int>(base::Time::kMillisecondsPerSecond *
-                                          kListenerRefreshRatio1 * ttl_));
+      last_update_ + base::Milliseconds(base::Time::kMillisecondsPerSecond *
+                                        kListenerRefreshRatio1 * ttl_);
 
   base::Time next_refresh2 =
-      last_update_ +
-      base::Milliseconds(static_cast<int>(base::Time::kMillisecondsPerSecond *
-                                          kListenerRefreshRatio2 * ttl_));
+      last_update_ + base::Milliseconds(base::Time::kMillisecondsPerSecond *
+                                        kListenerRefreshRatio2 * ttl_);
 
   base::SingleThreadTaskRunner::GetCurrentDefault()->PostDelayedTask(
       FROM_HERE, next_refresh_.callback(), next_refresh1 - clock_->Now());
